+++ src/extension/ExtensionContextHandler.ts
@@ -43,6 +43,8 @@ import { RecommendationCore } from '@redhat-developer/vscode-extension-proposals
 import { WhatsNewPanel } from './WhatsNewPanel';
 import { satisfies } from 'compare-versions';
 import { StepsOnSaveManager } from '../helpers/StepsOnSaveManager';
+import { CamelRunSourceDirJBangTask } from '../tasks/CamelRunSourceDirJBangTask';
+import { Folder } from '../views/integrationTreeItems/Folder';
 
 export class ExtensionContextHandler {
 	protected kieEditorStore: KogitoVsCode.VsCodeKieEditorStore;
@@ -325,6 +327,53 @@ export class ExtensionContextHandler {
 				await this.sendCommandTrackingEvent(INTEGRATIONS_RUN_COMMAND_ID);
 			}),
 		);
+	}
+
+	public registerRunSourceDirCommands(portManager: PortManager) {
+		const INTEGRATIONS_RUN_FOLDER_COMMAND_ID: string = 'kaoto.integrations.run.folder';
+		const INTEGRATIONS_RUN_WORKSPACE_COMMAND_ID: string = 'kaoto.integrations.run.workspace';
+		const INTEGRATIONS_RUN_ALL_WORKSPACES_COMMAND_ID: string = 'kaoto.integrations.run.all.workspaces';
+
+		const runSourceDirTask = async (folderPath: string) => {
+			const port = await portManager.allocatePort();
+			const runTask = await CamelRunSourceDirJBangTask.create(folderPath, port);
+			await runTask.execute();
+		};
+
+		const runFolderCommand = vscode.commands.registerCommand(INTEGRATIONS_RUN_FOLDER_COMMAND_ID, async (folder: Folder) => {
+			await runSourceDirTask(folder.folderUri.fsPath);
+			await this.sendCommandTrackingEvent(INTEGRATIONS_RUN_FOLDER_COMMAND_ID);
+		});
+
+		const folders = vscode.workspace.workspaceFolders;
+		const commandId = folders && folders.length > 1 ? INTEGRATIONS_RUN_ALL_WORKSPACES_COMMAND_ID : INTEGRATIONS_RUN_WORKSPACE_COMMAND_ID;
+		const runAllCommand = vscode.commands.registerCommand(commandId, async () => {
+			if (!folders) {
+				return;
+			}
+			for (const folder of folders) {
+				await runSourceDirTask(folder.uri.fsPath);
+			}
+
+			const storageKey = 'kaoto.showRunAllFoldersMessage';
+			let showInfoMessage = this.context.globalState.get<boolean>(storageKey, true);
+
+			if (showInfoMessage) {
+				const doNotShowAgain = "Don't show again";
+				const ok = 'OK';
+				const result = await vscode.window.showInformationMessage(
+					'You are running multiple workspaces. Each workspace will be run in a separate terminal.',
+					ok,
+					doNotShowAgain,
+				);
+				if (result === doNotShowAgain) {
+					await this.context.globalState.update(storageKey, false);
+				}
+			}
+			await this.sendCommandTrackingEvent(commandId);
+		});
+
+		this.context.subscriptions.push(runFolderCommand, runAllCommand);
 	}
 
 	public registerKubernetesRunCommands() {